Restructure test classes
