+++ src/test/java/teammates/test/BaseTestCaseWithLocalPostgresSqlDatabaseAccess.java
@@ -4,6 +4,9 @@ import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import org.hibernate.SessionFactory;
 import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
 import teammates.common.datatransfer.sqlattributes.CourseAttributes;
 import teammates.logic.api.LogicNewExtension;
 
@@ -12,6 +15,21 @@ import javax.sql.DataSource;
 public abstract class BaseTestCaseWithLocalPostgresSqlDatabaseAccess extends BaseTestCaseWithPostgresSqlDatabaseAccess {
     private static SessionFactory sessionFactory;
     private final LogicNewExtension logic = new LogicNewExtension();
+    protected static PostgreSQLContainer pgsql;
+
+    @BeforeClass
+    public static void startContainer() {
+        pgsql = new PostgreSQLContainer<>("postgres:14.4")
+                .withDatabaseName("pgsql-test")
+                .withUsername("pgsql-test")
+                .withPassword("pgsql-test");
+        pgsql.start();
+    }
+
+    @AfterClass
+    public static void closeContainer() {
+        pgsql.close();
+    }
 
     protected DataSource getDataSource(JdbcDatabaseContainer<?> container) {
         HikariConfig hikariConfig = new HikariConfig();
